pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pharmatrials-env"
version = "1.0.0"
description = "OpenEnv: Clinical-Trial Document Intelligence Environment"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [{name = "PharmaTrials-Env Contributors"}]
dependencies = [
    "pydantic>=2.5,<3.0",
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.29,<1.0",
    "jinja2>=3.1,<4.0",
    "openai>=1.30,<2.0",
    "pyyaml>=6.0,<7.0",
    "faker>=24.0,<25.0",
    "numpy>=1.26,<3.0",
    "rapidfuzz>=3.6,<4.0",
    "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
openenv = "pharmatrials_env.openenv_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pharmatrials_env*"]

[tool.setuptools.package-data]
pharmatrials_env = [
  "documents/templates/*.j2",
  "documents/vocabulary/*.json",
]

[tool.ruff]
exclude = [".venv", "__pycache__", "build", "dist"]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.vulture]
min_confidence = 80
paths = ["pharmatrials_env"]
